# [Enhancement] rank window function optimization, support row_number (3)