--- a/fe/fe-core/src/main/java/com/starrocks/analysis/SortInfo.java
+++ b/fe/fe-core/src/main/java/com/starrocks/analysis/SortInfo.java
@@ -52,6 +52,8 @@ public class SortInfo {
     // TODO: rethink this when we have a better cost model.
     private static final float SORT_MATERIALIZATION_COST_THRESHOLD = Expr.FUNCTION_CALL_COST;
 
+    // Only used in local partition topn
+    private List<Expr> partitionExprs_;
     private List<Expr> orderingExprs_;
     private final List<Boolean> isAscOrder_;
     // True if "NULLS FIRST", false if "NULLS LAST", null if not specified.
@@ -66,10 +68,11 @@ public class SortInfo {
     // sortTupleDesc_.
     private List<Expr> sortTupleSlotExprs_;
 
-    public SortInfo(List<Expr> orderingExprs, List<Boolean> isAscOrder,
+    public SortInfo(List<Expr> partitionExprs, List<Expr> orderingExprs, List<Boolean> isAscOrder,
                     List<Boolean> nullsFirstParams) {
         Preconditions.checkArgument(orderingExprs.size() == isAscOrder.size());
         Preconditions.checkArgument(orderingExprs.size() == nullsFirstParams.size());
+        partitionExprs_ = partitionExprs;
         orderingExprs_ = orderingExprs;
         isAscOrder_ = isAscOrder;
         nullsFirstParams_ = nullsFirstParams;
@@ -80,6 +83,7 @@ public class SortInfo {
      * C'tor for cloning.
      */
     private SortInfo(SortInfo other) {
+        partitionExprs_ = Expr.cloneList(other.partitionExprs_);
         orderingExprs_ = Expr.cloneList(other.orderingExprs_);
         isAscOrder_ = Lists.newArrayList(other.isAscOrder_);
         nullsFirstParams_ = Lists.newArrayList(other.nullsFirstParams_);
@@ -104,6 +108,10 @@ public class SortInfo {
             SlotDescriptor slotDesc = sortTupleDesc_.getSlots().get(i);
             slotDesc.setSourceExpr(sortTupleSlotExprs_.get(i));
         }
+    }
+
+    public List<Expr> getPartitionExprs() {
+        return partitionExprs_;
     }
 
     public List<Expr> getOrderingExprs() {