… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/Operator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/logical/LogicalTopNOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalTopNOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/scalar/ScalarOperator.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/AddDecodeNodeForDictStringRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rewrite/ExchangeSortToMergeRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/RuleType.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/implementation/TopNImplementationRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateWindowRankRule.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/transformer/QueryTransformer.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/transformer/WindowTransformer.java
… fe/fe-core/src/main/java/com/starrocks/sql/plan/PlanFragmentBuilder.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/WindowTest.java
… gensrc/thrift/PlanNodes.thrift