@liuyehcf 2022-05-18
run starrocks_fe_unittest