@wanpengfei-git 2022-05-23
#### [FE PR Coverage check]

:heart_eyes: **pass** : 127 / 141 (90.07%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/analysis/SortInfo.java|2|3|66.67%|
|:large_blue_circle:|com/starrocks/sql/optimizer/transformer/QueryTransformer.java|11|14|78.57%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/logical/LogicalTopNOperator.java|19|23|82.61%|
|:large_blue_circle:|com/starrocks/planner/SortNode.java|17|19|89.47%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateWindowRankRule.java|49|53|92.45%|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/Operator.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/TopNImplementationRule.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|12|12|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/ExchangeSortToMergeRule.java|3|3|100.00%|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|7|7|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/RuleType.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalTopNOperator.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/analysis/QueryStmt.java|1|1|100.00%|