@liuyehcf 2022-05-23
run starrocks_fe_unittest