@wanpengfei-git 2022-05-23
#### [FE PR Coverage check]

:heart_eyes: **pass** : 144 / 161 (89.44%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/analysis/SortInfo.java|4|6|66.67%|
|:large_blue_circle:|com/starrocks/sql/optimizer/transformer/QueryTransformer.java|11|14|78.57%|
|:large_blue_circle:|com/starrocks/planner/SortNode.java|20|24|83.33%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/logical/LogicalTopNOperator.java|23|27|85.19%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateWindowRankRule.java|52|56|92.86%|
|:large_blue_circle:|com/starrocks/sql/optimizer/Optimizer.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/Operator.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/implementation/TopNImplementationRule.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/base/SortProperty.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|12|12|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rewrite/ExchangeSortToMergeRule.java|4|4|100.00%|
|:large_blue_circle:|com/starrocks/planner/ExchangeNode.java|7|7|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/rule/RuleType.java|1|1|100.00%|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalTopNOperator.java|4|4|100.00%|
|:large_blue_circle:|com/starrocks/analysis/QueryStmt.java|1|1|100.00%|